backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-dotenv
google-generativeai
python-jose[cryptography]
psycopg2-binary